[common] Moved buffer.hpp into device_buffer.hpp
